Tested some commands. Added controls.
Tested and fixed the go commands, and implemented the cw and ccw command so that they rotate the drone and it moves according to how it is rotated. The drones do not rotate by themselves.
The drone won't move until it takes off. The space key takes off or lands the drone, then the arrow keys move the drone in their direction, the l key rotate the drone 45 degrees clockwise, and the j key rotates the drone 45 degrees counter clockwise. The z key prints the drone's state, and the q key allows u to enter any command u want (The program stops while u are entering the command). The delete key triggers the emergency command. 

Next milestone is to test all the commands and to rotate the drones visually.